+++ order/src/test/java/com/food/order/PaymentCancelTest.java
@@ -11,8 +11,8 @@ import com.food.order.stubrepository.StubPaymentLogService;
 import com.food.order.stubrepository.StubPaymentService;
 import com.food.order.stubrepository.StubPointService;
 import com.food.order.temp.DefaultPayService;
-import com.food.order.temp.InvalidPaymentActionTypeException;
-import com.food.order.temp.NotFoundPaymentException;
+import com.food.order.error.InvalidPaymentActionTypeException;
+import com.food.order.error.NotFoundPaymentException;
 import com.food.order.temp.PayService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
