+++ order/src/test/java/com/food/order/temp/DefaultPayService.java
@@ -14,10 +14,7 @@ import com.food.common.user.business.external.PointService;
 import com.food.common.user.business.external.model.PointCollectRequest;
 import com.food.common.user.business.external.model.PointUseRequest;
 import com.food.common.user.business.external.model.RequestUser;
-import com.food.order.error.DuplicatedPaymentException;
-import com.food.order.error.InvalidPaymentException;
-import com.food.order.error.NotFoundOrderException;
-import com.food.order.error.PaymentErrors;
+import com.food.order.error.*;
 import com.food.order.presentation.dto.request.PaymentDoRequest;
 import lombok.RequiredArgsConstructor;
